merge origin updates

+++ test/test_timeline.py
@@ -44,7 +44,7 @@ class TimelineTests(unittest.TestCase):
                 hvd.init()
 
                 # Perform a simple allreduce operation
-                hvd.allreduce(torch.tensor([1, 2, 3]), name='test_allreduce')
+                hvd.allreduce(torch.tensor([1, 2, 3], dtype=torch.float32), name='test_allreduce')
 
                 # Wait for it to register in the timeline.
                 time.sleep(0.1)
